requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Redis
redis==5.0.1
hiredis==2.2.3

# Kafka
kafka-python==2.0.2

# WebRTC & Media
websockets==12.0
av==12.0.0  # Updated from 9.2.0 - fixes Cython compilation errors

# Machine Learning
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.1.0
onnxruntime==1.16.3 # use 1.20.1
# onnxruntime-gpu==1.16.3  # Uncomment for GPU support

# Face recognition (InsightFace)
# Note: Models need to be downloaded separately
insightface==0.7.3  # Uncomment when models are available

# Vector database (pgvector for PostgreSQL)
pgvector==0.2.4  # PostgreSQL vector extension

# Image processing
scikit-image==0.22.0

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4